Don't close the menu during asset loading.
Back in the days it might have been necessary as the
game's own interrupt routines needed to be disabled during loading.
But now that we use OS interrupts and blitter queues, this is
no longer necessary.

This way we even get to see a nice animation while the assets are
being loaded.