Try fixing ChangeScreenBuffer again
The old code made the assumption that the ChangeScreenBuffer
will always send messages, regardless if it failed or not.
In failure case, we assumed we would need to wait for the "display safe"
message before being able to safely change screens.

But it looks like the messages are only sent out in success case.
In case of failure, no messages will be produced and we would get stuck
in the next frame waiting for the prior frame to finish display.

Fix by notifying the next frame if the call to ChangeScreenBuffer
was successful in the prior frame, thus we can expect a message to arrive.