Mark the VBLank interrupt exits
The VBlank interrupt exits in multiple places. Mark them as such.

The VBL does multiple things:
1) Test keyboard input and turn that into action.
Having the input sampling be independent of the main render loop makes
the game not miss input when the rendering is slow. This makes AB3D feel rather
responsive even when at low framerates.
The actual sampling of the keyboard is done in the Keyboard interrupt.

2) Serial communication in multiplyer games

3) Time floor damage(?) and set damage palette

4) Alien animation, every 5th frame , thus animation is supposed to run at 10fps
   This will also cause Aliens to make noises tied to animations

5) Sound mixing.